Makes auth token expire only if not used after a while
We changed auth tokens so they would expire if a new session was created - as they were always meant to - in eedfd00c4b4de09d765e117101507e589f0ee589.

This of course broke everyone who depended on the tokens never expiring - particularly for API based programs.

Until we can provide an alternative (see Allow users to mint JWTs for scripts #727) we've made token stable _while they continue to be used_. Any session interaction will prolong the life of the token. This should make it so anyone using a token regularly - e.g. for scripted access - will be able to do so until their activity stops.

Fixes Token expiration too harsh #728
